+++ packages/kysely/src/db.ts
@@ -1,6 +1,5 @@
 import { z } from "@hono/zod-openapi";
 import { SqlBranding } from "./branding/Branding";
-import { SqlUser } from "./users/User";
 import {
   applicationSchema,
   Code,
@@ -14,6 +13,7 @@ import {
   SigningKey,
   Tenant,
   themeSchema,
+  userSchema,
 } from "@authhero/adapter-interfaces";
 import { SqlTicket } from "./tickets/Ticket";
 import { SqlLog } from "./logs/Log";
@@ -23,17 +23,13 @@ const sqlThemeSchema = flattenSchema(themeSchema).extend({
   tenant_id: z.string(),
 });
 
-type SqlTheme = z.infer<typeof sqlThemeSchema>;
-
 const sqlLoginSchema = flattenSchema(loginSchema).extend({
   tenant_id: z.string(),
 });
-type SqlLogin = z.infer<typeof sqlLoginSchema>;
 
 const sqlConnectionSchema = flattenSchema(connectionSchema).extend({
   tenant_id: z.string(),
 });
-type SqlConnection = z.infer<typeof sqlConnectionSchema>;
 
 const sqlApplicationSchema = z.object({
   ...applicationSchema.shape,
@@ -56,21 +52,30 @@ const sqlPromptSettingSchema = z.object({
   tenant_id: z.string(),
 });
 
+const sqlUserSchema = z.object({
+  ...userSchema.shape,
+  email_verified: z.number(),
+  is_social: z.number(),
+  app_metadata: z.string(),
+  user_metadata: z.string(),
+  tenant_id: z.string(),
+});
+
 export interface Database {
   applications: z.infer<typeof sqlApplicationSchema>;
   branding: SqlBranding;
   codes: Code & { tenant_id: string };
-  connections: SqlConnection;
+  connections: z.infer<typeof sqlConnectionSchema>;
   domains: Domain & { tenant_id: string };
   hooks: Hook & { tenant_id: string };
   keys: SigningKey & { created_at: string };
-  logins: SqlLogin;
+  logins: z.infer<typeof sqlLoginSchema>;
   logs: SqlLog;
   passwords: Password & { tenant_id: string };
   prompt_settings: z.infer<typeof sqlPromptSettingSchema>;
-  users: SqlUser;
+  users: z.infer<typeof sqlUserSchema>;
   sessions: Session & { tenant_id: string };
   tenants: Tenant;
-  themes: SqlTheme;
+  themes: z.infer<typeof sqlThemeSchema>;
   tickets: SqlTicket;
 }