+++ packages/kysely/src/users/list.ts
@@ -53,6 +53,8 @@ export function listUsers(db: Kysely<Database>) {
         ...user,
         email_verified: user.email_verified === 1,
         is_social: user.is_social === 1,
+        app_metadata: JSON.parse(user.app_metadata),
+        user_metadata: JSON.parse(user.user_metadata),
         identities: [
           {
             connection: user.connection,